requirements.txt
Flask
Flask-Cors
Flask-Session
python-dotenv